+++ next.config.js
@@ -1,6 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: false, // react-form-validates 高阶函数会报warning。可参考 './warning.png'
-}
-
-module.exports = nextConfig